server/whisper-worker/requirements.txt
faster-whisper==1.0.1
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.3
numpy==1.24.3
torch>=2.0.0
torchaudio>=2.0.0
prometheus-client==0.19.0
structlog==24.1.0
watchdog==3.0.0